+++ slasher/db/kv/BUILD.bazel
@@ -7,6 +7,7 @@ go_library(
         "attester_slashings.go",
         "block_header.go",
         "chain_data.go",
+        "epoch_store.go",
         "indexed_attestations.go",
         "kv.go",
         "proposer_slashings.go",
@@ -43,6 +44,7 @@ go_test(
         "benchmark_test.go",
         "block_header_test.go",
         "chain_data_test.go",
+        "epoch_store_test.go",
         "indexed_attestations_test.go",
         "kv_test.go",
         "proposer_slashings_test.go",
