+++ slasher/db/kv/spanner_new.go
@@ -2,63 +2,11 @@ package kv
 
 import (
 	"context"
-	"errors"
 
 	"github.com/prysmaticlabs/prysm/shared/bytesutil"
-	"github.com/prysmaticlabs/prysm/slasher/detection/attestations/types"
 	bolt "go.etcd.io/bbolt"
 	"go.opencensus.io/trace"
 )
-
-var errWrongSize = errors.New("wrong data length for min max span byte array")
-var highestObservedValidatorIdx uint64
-
-// GetValidatorSpan unmarshal a span from an encoded, flattened array.
-func (es EpochStore) GetValidatorSpan(ctx context.Context, idx uint64) (types.Span, error) {
-	r := types.Span{}
-	if len(es)%spannerEncodedLength != 0 {
-		return r, errWrongSize
-	}
-	origLength := uint64(len(es)) / spannerEncodedLength
-	requestedLength := idx + 1
-	if origLength < requestedLength {
-		return r, nil
-	}
-	cursor := idx * spannerEncodedLength
-	r.MinSpan = bytesutil.FromBytes2(es[cursor : cursor+2])
-	r.MaxSpan = bytesutil.FromBytes2(es[cursor+2 : cursor+4])
-	sigB := [2]byte{}
-	copy(sigB[:], es[cursor+4:cursor+6])
-	r.SigBytes = sigB
-	r.HasAttested = bytesutil.ToBool(es[cursor+6])
-	return r, nil
-}
-
-// SetValidatorSpan marshal a validator span into an encoded, flattened array.
-func (es EpochStore) SetValidatorSpan(ctx context.Context, idx uint64, newSpan types.Span) error {
-	if len(es)%spannerEncodedLength != 0 {
-		return errors.New("wrong data length for min max span byte array")
-	}
-	if highestObservedValidatorIdx < idx {
-		highestObservedValidatorIdx = idx
-	}
-	if len(es) == 0 {
-		requestedLength := highestObservedValidatorIdx*spannerEncodedLength + spannerEncodedLength
-		es = make([]byte, requestedLength, requestedLength)
-	}
-	cursor := idx * spannerEncodedLength
-	endCursor := cursor + spannerEncodedLength
-	spansLength := uint64(len(es))
-	if endCursor > spansLength {
-		diff := endCursor - spansLength
-		b := make([]byte, diff, diff)
-		es = append(es, b...)
-	}
-	enc := marshalSpan(newSpan)
-	copy(es[cursor:], enc)
-
-	return nil
-}
 
 // EpochSpans accepts epoch and returns the corresponding spans byte array
 // for slashing detection.
@@ -90,7 +38,7 @@ func (db *Store) SaveEpochSpans(ctx context.Context, epoch uint64, es EpochStore
 	defer span.End()
 
 	if len(es)%spannerEncodedLength != 0 {
-		return errWrongSize
+		return ErrWrongSize
 	}
 	return db.update(func(tx *bolt.Tx) error {
 		b, err := tx.CreateBucketIfNotExists(validatorsMinMaxSpanBucketNew)
